MT5/api/pyproject.toml
[project]
name = "app"
version = "0.2.1"
#description = "Add your description here"
readme = "README.md"
requires-python = "==3.14.3"
dependencies = [
    "apscheduler>=3.11.2",
    "fastapi>=0.125.0",
    "metatrader5>=5.0; sys_platform == 'win32'",
    #"numpy>=2.0.2",
    "pandas>=2.3.3",
    "prometheus-fastapi-instrumentator>=7.1.0",
    "pydantic-settings>=2.13",
    #"pyparsing>=3.3.2",
    "python-dotenv>=1.2.1",
    #    "python-multipart>=0.0.22",
    "pytz>=2025.2",
    "pyyaml>=6.0.3",
    #    "requests>=2.32.5",
    "sqlalchemy>=2.0.48",
    "sqlmodel>=0.0.30",
    "uvicorn>=0.39.0",
]

[dependency-groups]
dev = [
    'pytest',
    'httpx'
]

[tool.pytest.ini_options]
pythonpath = ["."]